feat: :sparkles: Introdcue SelectTickets & Modal
Enter what will be the page to select tickets, it is coded to only access from checkout. Modal is created to enter product details Filters are adapted to the structure of the products in Raffles. The old way of bringing products is eliminated Cart status is saved in localstorage Add to cart logic is passed directly to the ButtonAddTicket button. The function that is responsible for saving in the cart is factored to improve performance.

Build project